Renamed "DLGPS" to "DLLocation"
